mobile/android/tests/browser/chrome/test_selectoraddtab.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=1216047
-->
<head>
  <meta charset="utf-8">
  <title>Test for Bug 1216047</title>
  <script type="application/javascript" src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
  <script type="application/javascript" src="chrome://mochikit/content/tests/SimpleTest/SpawnTask.js"></script>
  <link rel="stylesheet" type="text/css" href="chrome://global/skin"/>
  <link rel="stylesheet" type="text/css" href="chrome://mochikit/content/tests/SimpleTest/test.css"/>
  <script type="application/javascript" src="head.js"></script>
  <script type="application/javascript;version=1.7">

  "use strict";

  const { classes: Cc, interfaces: Ci, utils: Cu } = Components;

  Cu.import("resource://gre/modules/Services.jsm");
  Cu.import("resource://gre/modules/Messaging.jsm");
  Cu.import("resource://gre/modules/Task.jsm");

  // The chrome window
  let chromeWin;

  // Track the <browser>s where the tests are happening
  let browserBlank;
  let browserTest;

  const kTestPage = "http://mochi.test:8888/chrome/mobile/android/tests/browser/chrome/basic_article.html";

  add_task(function* test_selectOrAdd() {
    chromeWin = Services.wm.getMostRecentWindow("navigator:browser");
    let BrowserApp = chromeWin.BrowserApp;

    SimpleTest.registerCleanupFunction(function() {
      BrowserApp.closeTab(BrowserApp.getTabForBrowser(browserBlank));
      BrowserApp.closeTab(BrowserApp.getTabForBrowser(browserTest));
    });

    // Add a new tab with a blank page
    browserBlank = BrowserApp.addTab("about:blank", { selected: true, parentId: BrowserApp.selectedTab.id }).browser;

    // Now, let's force the target browser to be added
    browserTest = BrowserApp.selectOrAddTab(kTestPage, { selected: true, parentId: BrowserApp.selectedTab.id }).browser;
    yield promiseBrowserEvent(browserTest, "DOMContentLoaded");

    // Check that basic_article is now selected
    is(BrowserApp.selectedBrowser, browserTest, "Target browser is selected after being added.");

    // Switch back to about:blank
    BrowserApp.selectTab(BrowserApp.getTabForBrowser(browserBlank));
    yield promiseTabEvent(BrowserApp.deck, "TabSelect");

    // Check that about:blank is selected
    is(BrowserApp.selectedBrowser, browserBlank, "about:blank is selected.");

    // Use selectOrAddTab to select the existing tab
    BrowserApp.selectOrAddTab(kTestPage, { selected: true, parentId: BrowserApp.selectedTab.id }).browser;
    yield promiseTabEvent(BrowserApp.deck, "TabSelect");

    // Check that basic_article is now selected
    is(BrowserApp.selectedBrowser, browserTest, "Target browser is selected.");

    // Switch back to about:blank
    BrowserApp.selectTab(BrowserApp.getTabForBrowser(browserBlank));
    yield promiseTabEvent(BrowserApp.deck, "TabSelect");

    // Check that about:blank is selected
    is(BrowserApp.selectedBrowser, browserBlank, "about:blank is selected.");

    // Use selectOrAddTab to select the existing tab using the startsWith flag
    BrowserApp.selectOrAddTab(kTestPage, { selected: true, parentId: BrowserApp.selectedTab.id }, { startsWith: kTestPage }).browser;
    yield promiseTabEvent(BrowserApp.deck, "TabSelect");

    // Check that basic_article is now selected
    is(BrowserApp.selectedBrowser, browserTest, "Target browser is selected.");
  });

  </script>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=1216047">Mozilla Bug 1216047</a>
<p id="display"></p>
<div id="content" style="display: none">

</div>
<pre id="test">
</pre>
</body>
</html>
